Broke temperature out into its own component.
